requirements-dev.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.24
django-cors-headers==3.2.1
django-filter==2.2.0
django-taggit==1.2.0
django-taggit-serializer==0.1.7
django-webpack-loader==0.6.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.8
itypes==1.1.0
Jinja2==2.11.1
Markdown==3.2
MarkupSafe==1.1.1
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
